message class refactor
